+++ src/plugins/vis_builder/public/embeddable/vis_builder_embeddable_factory.tsx
@@ -84,6 +84,7 @@ export class VisBuilderEmbeddableFactory
         getTimeFilter(),
         {
           savedVis,
+          savedObject,
           editUrl,
           editPath,
           editable: true,